+++ http/resource.go
@@ -118,9 +118,7 @@ func resourceDeleteHandler(fileCache FileCache) handleFunc {
 			return errToStatus(err), err
 		}
 
-		err = d.RunHook(func() error {
-			return d.user.Fs.RemoveAll(r.URL.Path)
-		}, "delete", r.URL.Path, "", d.user)
+		err = d.user.Fs.RemoveAll(r.URL.Path)
 
 		if err != nil {
 			return errToStatus(err), err
@@ -138,9 +136,7 @@ func resourcePostHandler(fileCache FileCache) handleFunc {
 
 		// Directories creation on POST.
 		if strings.HasSuffix(r.URL.Path, "/") {
-			err := d.RunHook(func() error {
-				return d.user.Fs.MkdirAll(r.URL.Path, d.settings.DirMode)
-			}, "upload", r.URL.Path, "", d.user)
+			err := d.user.Fs.MkdirAll(r.URL.Path, d.settings.DirMode)
 			return errToStatus(err), err
 		}
 
@@ -168,16 +164,14 @@ func resourcePostHandler(fileCache FileCache) handleFunc {
 			}
 		}
 
-		err = d.RunHook(func() error {
-			info, writeErr := writeFile(d.user.Fs, r.URL.Path, r.Body, d.settings.FileMode, d.settings.DirMode)
-			if writeErr != nil {
-				return writeErr
-			}
-
+		info, writeErr := writeFile(d.user.Fs, r.URL.Path, r.Body, d.settings.FileMode, d.settings.DirMode)
+		if writeErr != nil {
+			err = writeErr
+		} else {
 			etag := fmt.Sprintf(`"%x%x"`, info.ModTime().UnixNano(), info.Size())
 			w.Header().Set("ETag", etag)
-			return nil
-		}, "upload", r.URL.Path, "", d.user)
+			err = nil
+		}
 
 		if err != nil {
 			_ = d.user.Fs.RemoveAll(r.URL.Path)
@@ -205,16 +199,14 @@ var resourcePutHandler = withUser(func(w http.ResponseWriter, r *http.Request, d
 		return http.StatusNotFound, nil
 	}
 
-	err = d.RunHook(func() error {
-		info, writeErr := writeFile(d.user.Fs, r.URL.Path, r.Body, d.settings.FileMode, d.settings.DirMode)
-		if writeErr != nil {
-			return writeErr
-		}
-
+	info, writeErr := writeFile(d.user.Fs, r.URL.Path, r.Body, d.settings.FileMode, d.settings.DirMode)
+	if writeErr != nil {
+		err = writeErr
+	} else {
 		etag := fmt.Sprintf(`"%x%x"`, info.ModTime().UnixNano(), info.Size())
 		w.Header().Set("ETag", etag)
-		return nil
-	}, "save", r.URL.Path, "", d.user)
+		err = nil
+	}
 
 	return errToStatus(err), err
 })
@@ -267,9 +259,7 @@ func resourcePatchHandler(fileCache FileCache) handleFunc {
 			return errToStatus(err), err
 		}
 
-		err = d.RunHook(func() error {
-			return patchAction(r.Context(), action, src, dst, d, fileCache)
-		}, action, src, dst, d.user)
+		err = patchAction(r.Context(), action, src, dst, d, fileCache)
 
 		return errToStatus(err), err
 	})